requirements.txt
fastapi==0.70.0
pydantic==1.8.2
python-dotenv==0.19.2
starlette==0.17.1
telethon==1.23.0
uvicorn==0.15.0
